+++ packages/connectors/src/walletConnect.ts
@@ -13,7 +13,7 @@ import {
 } from 'viem'
 
 import { Connector } from './base'
-import { StorageStoreData } from './types'
+import type { StorageStoreData, WalletClient } from './types'
 
 type WalletConnectOptions = {
   /**
@@ -67,9 +67,14 @@ type WalletConnectOptions = {
   showQrModal?: EthereumProviderOptions['showQrModal']
   /**
    * Options of QR code modal.
-   * @link https://docs.walletconnect.com/2.0/web3modal/options
+   * @link https://docs.walletconnect.com/2.0/web/walletConnectModal/modal/options
    */
   qrModalOptions?: EthereumProviderOptions['qrModalOptions']
+  /**
+   * Option to override default relay url.
+   * @link https://docs.walletconnect.com/2.0/web/providers/ethereum
+   */
+  relayUrl?: string
 }
 
 type ConnectConfig = {
@@ -186,7 +191,9 @@ export class WalletConnectConnector extends Connector<
     return this.#provider!
   }
 
-  async getWalletClient({ chainId }: { chainId?: number } = {}) {
+  async getWalletClient({
+    chainId,
+  }: { chainId?: number } = {}): Promise<WalletClient> {
     const [provider, account] = await Promise.all([
       this.getProvider({ chainId }),
       this.getAccount(),
@@ -289,6 +296,7 @@ export class WalletConnectConnector extends Connector<
         showQrModal = true,
         qrModalOptions,
         metadata,
+        relayUrl,
       } = this.options
       this.#provider = await EthereumProvider.init({
         showQrModal,
@@ -305,6 +313,7 @@ export class WalletConnectConnector extends Connector<
           ]),
         ),
         metadata,
+        relayUrl,
       })
     }
   }